Added resources for Period, PossibleSchedule and Professor. Updated attributes on PossibleSchedule

+++ speedplanner/src/main/java/com/opensource/speedplanner/model/Professor.java
@@ -21,7 +21,7 @@ public class Professor { //Agregar anotaciones
     private Long id;
 
     @Column(unique=true)
-    private Long code;
+    private int code;
 
     @Column(unique=true)
     @NotBlank
